move `certificate.go` in it's own package and make functions public
